Fix pretty printing of cons
Right now, any expressions with cons don't print back correctly, due to incorrect parentheses. Part of this is because :: is being parsed as a binOp, but the prettyPrinter is looking for :: as a special case of Data.
This change adds the check from Data to BinOp, so that cons expressions print without parentheses.